habitat_integration/requirements.txt
# Python dependencies for Habitat integration with NoMaD

# Core dependencies (should already be in nomad_train environment)
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
pillow>=9.0.0
pyyaml>=5.4.1
matplotlib>=3.5.0

# Diffusion model dependencies
diffusers==0.11.1
huggingface_hub<0.19.0  # Compatible with diffusers 0.11.1

# Vision encoder
efficientnet-pytorch>=0.7.1

# Habitat simulator and lab
habitat-sim>=0.2.3
habitat-lab>=0.2.3

# Optional but recommended
tqdm>=4.64.0
